Database & Query Updates
Added an integer field to the DiscreteVisitor table to track the most
recently frame when visitor rows are updated.

The add_processed_video function was modified to set the
pollinators_processed boolean field to true for records with a given
video filename. A print function was added to communicate this action
to the user.

The get_processed_videos function now includes an optional parameter
depending if the videos being requested are focused on frame or
pollinator processing.

Signed-off-by: Avery Uslaner <avery.uslaner@redbutte.utah.edu>