Data Storage Enhancements
Added date field to DiscreteVisitor model.

Added function to request the datetime stamp of a given video frame from
the database.

Altered the function controlling addition or updating of discrete
visitor records so that the powerpoint slide and notes fields are not
considered for default values. Also added the date field to the function
parameters.

Signed-off-by: Avery Uslaner <avery.uslaner@redbutte.utah.edu>